feat: move to shared directory
