+++ cypress/component/ProgressReport/SubmitProgressReport.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {mount} from 'cypress/react';
-import SubmitProgressReport from '../../../src/pages/progress_report/SubmitProgressReport';
+import SubmitProgressReport from 'src/pages/progress_reports/SubmitProgressReport';
 import {StackdriverReporter} from '../../../src/libs/stackdriverReporter';
 import '../../../src/index.css';
 import '../../../src/styles/buttons.css';
